Merchant_X_Consumer/requirements.txt
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1708969572489/work
asgiref==3.10.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.12.3
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1711936407380/work
Brotli @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/brotli-split_1699473013692/work
bs4==0.0.2
certifi==2025.6.15
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
colorama==0.4.6
comm==0.2.2
conda @ file:///D:/bld/conda_1711446014277/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1706566000184/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
cryptography==42.0.5
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1704321475663/work
Django==5.2.7
django-allauth==0.61.1
django-extensions==4.1
django-filter==25.2
django-silk==5.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
executing==2.2.0
gprof2dot==2025.4.14
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
inflection==0.5.1
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
libmambapy @ file:///C:/b/abs_7dmjutgtwb/croot/mamba-split_1712091963973/work/libmambapy
lxml==5.2.1
matplotlib-inline==0.1.7
menuinst @ file:///C:/b/abs_099kybla52/croot/menuinst_1706732987063/work
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.2
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1706116770704/work
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycosat @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pycosat_1699482932804/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
Pygments==2.19.2
PyJWT==2.8.0
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2024.1
pywin32==310
PyYAML==6.0.3
pyzmq==27.0.0
referencing==0.37.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==2.0.0
rpds-py==0.27.1
ruamel.yaml @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/ruamel.yaml_1699483184324/work
selenium==4.34.0
setuptools==68.2.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.5.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1707598593068/work
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1694154605758/work
typing_extensions==4.14.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wget==3.2
wheel==0.41.2
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wsproto==1.2.0
zstandard==0.19.0
pytest==9.0.2
pytest-django==4.11.1
factory_boy==3.3.3